Snapshot of work so far
I managed to get up to a "working" (untested, but it compiles!)
version of does_fmt_match_args.

I want to change how I do some things in order to make this a little
more flexible, so I'm going to be starting from the bottom up again,
but this is worth saving in version history.