Revisions to ensure the trick I'm using will actually work
The trick in question is that we can use certain structs as arguments
to functions and the argument layout will be the same as if the struct's
fields were used individually as arguments in the same order.

This is not true for arbitrary ABIs, but (within certain limits!) *is*
true on the (architecture, ABI) pairs I actually care about right now.