Major re-jiggering of argument-list types
Basically, the old scheme, which I *thought* would work on several
architecture's SysV ABIs, does not in fact work on x86-64. This is
what I get for trying to be clever. As such, a rethink was in order.

The downsides of the new scheme are that (1) arguments corresponding
to '*' parts of conversion specifications don't get visually grouped with
their corresponding value arguments in tuples and (2) for now I don't
know how to safely support `&str` (as opposed to `NullString` or `&CStr`)
arguments. The upsides are that (1) there's significantly less cleverness
involved, and as a result, (2) this should Just Work on any POSIX system,
regardless of ABI.

This also includes fixes for typos in test cases; I wasn't mentally
interpreting %f conversions correctly for my output values.