Removed array-index hack in favor of panic!
Prior to Rust 1.57.0, you could not use the `panic!` macro in a
`const fn`. Now, however, you can, at least to a limited extent. [1]
Prior to this commit, we used a hack around invalid array indices in
order to panic in `const fn`s. This commit switches over to `panic!`,
which is much more straightforward to understand.

[1] https://blog.rust-lang.org/2021/12/02/Rust-1.57.0.html#panic-in-const-contexts