Switched import of CStr from `std` to `core`
`CStr` should be available in `core` starting in Rust 1.64.0, so doing
this, along with a couple other things in this commit, allows us to get
rid of conditionally depending on `std` to get support for `CStr`
arguments. Indeed, by switching from implementing `PrintfArgument`
directly on `CStr` and `CString` to implementing it on `T: AsRef<CStr>`,
we don't even after to explicitly pull in `alloc` (except for tests).